Huge performance improvement
- Huge performance improvement:
Each call to any socket's read method adds a 50ms delay, to not have horrible speed, mt-buffer lib is used for buffering, however it turned out it wasn't really working that well and that for some reason each packet took a long time. The thing was just that serverSocket:pollState() called internet.read everytime (so 50ms delay per packet), it is now only called when buffer is empty (so when all packets got processed) which is completely fine for the purpose of the pollState() method.

- Bridge source: The source code of the "bridge" software is now released

- Player rotation is now theoritically supported (not tested yet!)

- Players can now load chunks and be kicked

- Custom terrain generation with perlin noise

- Ability to place blocks other than dirt (note only a few are supported, more to come with data generators)